Fix compilation on macos.
This also updates the testing to do more thorough testing accross
platforms, as that is creating just too many issues.
