+++ src/raw_socket.rs
@@ -15,8 +15,6 @@ use crate::{
     },
 };
 
-#[cfg(any(target_os = "macos", target_os = "freebsd"))]
-mod bsdlike;
 #[cfg(not(any(target_os = "macos", target_os = "freebsd", target_os = "linux")))]
 mod fallback;
 #[cfg(target_os = "freebsd")]
